Added different encounters to pythonProject.py
Added the Team Rocket and item encounters for the game. If Team Rocket is encountered, they will steal the first pokemon in the player's caught pokemon. That pokemon will return to the wild pokemon list.
Right now, the item encounter is +2 pokeballs.

Also added a chance mechanic so every time the player goes into the tall grass, it will randomly choose a pokemon, team rocket, or item encounter.