More updates on pythonProject.py
Randomized the item encounter so the player will receive either berries or pokeballs. Enhanced the berry throwing mechanic so they increase the chance of a pokemon being caught.
Added more text at the beginning of the game to add the player's name and to explain a bit more about how the game works