css and js edits to frontend project
CSS: Changed the layout of the cards to image on top, title and states on the bottom. Changed font-size of park titles.

JS: Changed user input to a drop box of states instead of a text input. Added spaces between states when the park is located in multiple states and added a function to add an ellipsis if the number of states was too long.